apis, utils 관련 모듈화 작업 중 1차 테스트 완료
