+++ popcornKiller/pk_apps/accounts/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class DiariesConfig(AppConfig):
+class AccountsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'apps.diaries'
+    name = 'pk_apps.accounts'
