+++ popcornKiller/pk_apps/box_office/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class AccountsConfig(AppConfig):
+class BoxofficeConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'apps.accounts'
+    name = 'pk_apps.box_office'
