+++ popcornKiller/pk_apps/box_office/templates/box_office/movie_detail.html
