+++ popcornKiller/pk_utils/api_utils.py
@@ -1,0 +1,10 @@
+import requests, json
+
+
+def fetch_api_data(url):
+    try:
+        response = requests.get(url)
+        response.raise_for_status()
+        return response.json()
+    except (requests.RequestException, json.JSONDecodeError):
+        return None
