+++ popcornKiller/pk_utils/general_utils.py
@@ -1,0 +1,5 @@
+from django.shortcuts import render
+
+
+def get_error_response(request, message):
+    return render(request, 'error.html', {'error': message})
