+++ popcornKiller/popcornKiller/urls.py
@@ -3,7 +3,7 @@ from django.urls import path, include
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('', include('apps.box_office.urls')),
-    path('accounts/', include('apps.accounts.urls')),
-    path('diaries/', include('apps.diaries.urls')),
+    path('', include('pk_apps.box_office.urls')),
+    path('accounts/', include('pk_apps.accounts.urls')),
+    path('diaries/', include('pk_apps.diaries.urls')),
 ]
